working

+++ lib/saveData.py
@@ -12,21 +12,25 @@ class DataSaver:
         if not  os.path.isdir(self.dir):
             os.makedirs(self.dir)
         
-        with open(filename + '.csv', 'w') as file:
-            file.write(self.divider.join(self.titles))
+        if not os.path.isfile(self.filename):        
+            with open(filename + '.csv', 'w') as file:
+                file.write(self.divider.join(self.titles))
+
+    def newLine(self, data, *rest):
+        row = []
+
+        for title in self.titles:
+            value = data[title]
+            if isinstance(value, int) or isinstance(value, float) or isinstance(value, numpy.int64) or isinstance(value, numpy.float32):
+                strValue = str(value)
+            else:
+                strValue =  ' '.join(str(x) for x in value)
+            row.append( strValue )
+
+        return self.divider.join(row)
 
     def add(self, data, *rest):
         # import pdb; pdb.set_trace()
         with open(self.filename, 'a') as file:
             file.write('\n')
-            row = []
-
-            for title in self.titles:
-                value = data[title]
-                if isinstance(value, int) or isinstance(value, float) or isinstance(value, numpy.int64) or isinstance(value, numpy.float32):
-                    strValue = str(value)
-                else:
-                    strValue =  ' '.join(str(x) for x in value)
-                
-                row.append( strValue )
-            file.write(self.divider.join(row))
+            file.write(self.newLine(data, *rest))
